requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.4.4
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
netifaces==0.10.6
pkg-resources==0.0.0
pycodestyle==2.5.0
python-engineio==3.10.0
python-socketio==4.3.1
six==1.12.0
typing-extensions==3.7.4
yarl==1.3.0
